+++ common/src/main/java/net/arna/jcraft/client/renderer/entity/HGNetRenderer.java
@@ -12,4 +12,9 @@ public class HGNetRenderer extends GeoEntityRenderer<HGNetEntity> {
         addRenderLayer(new HGNetGlowLayer(this));
         shadowRadius = 1.25f;
     }
+
+    @Override
+    public boolean shouldShowName(HGNetEntity animatable) {
+        return false;
+    }
 }