+++ common/src/main/java/net/arna/jcraft/client/renderer/entity/SandTornadoRenderer.java
@@ -9,12 +9,15 @@ import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.resources.ResourceLocation;
 import org.jetbrains.annotations.Nullable;
 
-
 public class SandTornadoRenderer extends GeoEntityRenderer<SandTornadoEntity> {
-
     @Override
     public RenderType getRenderType(SandTornadoEntity animatable, ResourceLocation texture, @Nullable MultiBufferSource bufferSource, float partialTick) {
         return RenderType.entityTranslucent(getTextureLocation(animatable));
+    }
+
+    @Override
+    public boolean shouldShowName(SandTornadoEntity animatable) {
+        return false;
     }
 
     public SandTornadoRenderer(EntityRendererProvider.Context renderManagerIn) {