+++ common/src/main/java/net/arna/jcraft/client/renderer/entity/projectiles/GeoProjectileRenderer.java
@@ -12,10 +12,14 @@ import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.util.Mth;
 import net.minecraft.world.entity.Entity;
 
-
 public class GeoProjectileRenderer<T extends Entity& GeoAnimatable> extends GeoEntityRenderer<T> {
     public GeoProjectileRenderer(EntityRendererProvider.Context renderManager, GeoModel<T> model) {
         super(renderManager, model);
+    }
+
+    @Override
+    public boolean shouldShowName(T animatable) {
+        return false;
     }
 
     @Override