+++ common/src/main/java/net/arna/jcraft/client/renderer/entity/stands/SunRenderer.java
@@ -1,26 +1,19 @@
 package net.arna.jcraft.client.renderer.entity.stands;
 
-import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
-import com.mojang.blaze3d.vertex.VertexMultiConsumer;
 import mod.azure.azurelib.cache.object.BakedGeoModel;
-import mod.azure.azurelib.core.animation.AnimationState;
 import mod.azure.azurelib.renderer.GeoEntityRenderer;
 import net.arna.jcraft.client.model.entity.TheSunModel;
 import net.arna.jcraft.client.renderer.entity.layer.SunGlowLayer;
 import net.arna.jcraft.common.entity.stand.TheSunEntity;
-import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Direction;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.Mth;
-import net.minecraft.world.entity.Pose;
 import org.jetbrains.annotations.Nullable;
-import org.joml.Matrix4f;
 
 public class SunRenderer extends GeoEntityRenderer<TheSunEntity> {
     public SunRenderer(EntityRendererProvider.Context renderManager) {
@@ -39,7 +32,6 @@ public class SunRenderer extends GeoEntityRenderer<TheSunEntity> {
     }
 
     //TODO: translucent layer that isn't layered over and has no shading
-
     @Override
     public RenderType getRenderType(TheSunEntity animatable, ResourceLocation texture, @Nullable MultiBufferSource bufferSource, float partialTick) {
         return RenderType.dragonExplosionAlpha(texture);
@@ -56,65 +48,10 @@ public class SunRenderer extends GeoEntityRenderer<TheSunEntity> {
         packedLight = 255;
 
         poseStack.pushPose();
-
         float scale = lerpScale(animatable, partialTick);
         poseStack.scale(scale, scale, scale);
 
-        this.modelRenderTranslations = new Matrix4f(poseStack.last().pose());
-
-        float lerpBodyRot = Mth.rotLerp(partialTick, animatable.yBodyRotO, animatable.yBodyRot);
-        float lerpHeadRot = Mth.rotLerp(partialTick, animatable.yHeadRotO, animatable.yHeadRot);
-        float netHeadYaw = lerpHeadRot - lerpBodyRot;
-
-        if (animatable.getPose() == Pose.SLEEPING) {
-            Direction bedDirection = animatable.getBedOrientation();
-
-            if (bedDirection != null) {
-                float eyePosOffset = animatable.getEyeHeight(Pose.STANDING) - 0.1F;
-
-                poseStack.translate(-bedDirection.getStepX() * eyePosOffset, 0, -bedDirection.getStepZ() * eyePosOffset);
-            }
-        }
-
-        float ageInTicks = animatable.tickCount + partialTick;
-        float limbSwingAmount = 0;
-        float limbSwing = 0;
-
-        applyRotations(animatable, poseStack, ageInTicks, lerpBodyRot, partialTick);
-
-        AnimationState<TheSunEntity> predicate = new AnimationState<TheSunEntity>(animatable, limbSwing, limbSwingAmount, partialTick, false);
-
-
-        this.model.setCustomAnimations(animatable, getInstanceId(animatable), predicate);
-
-        //poseStack.translate(0, 0.01f, 0);
-        RenderSystem.setShaderTexture(0, getTextureLocation(animatable));
-
-        var renderColor = getRenderColor(animatable, partialTick, packedLight);
-
-
-        if (!animatable.isInvisibleTo(Minecraft.getInstance().player)) {
-            VertexConsumer glintBuffer = bufferSource.getBuffer(RenderType.entityGlintDirect());
-            VertexConsumer translucentBuffer = bufferSource
-                    .getBuffer(RenderType.entityTranslucentCull(getTextureLocation(animatable)));
-
-            super.actuallyRender(
-                    poseStack,
-                    animatable,
-                    model,
-                    renderType,
-                    bufferSource,
-                    glintBuffer != translucentBuffer ? VertexMultiConsumer.create(glintBuffer, translucentBuffer) : null,
-                    isReRender,
-                    partialTick,
-                    packedLight,
-                    getPackedOverlay(animatable, 0), renderColor.getRed() / 255f,
-                    renderColor.getGreen() / 255f, renderColor.getBlue() / 255f,
-                    renderColor.getAlpha() / 255f
-            );
-        }
-
+        super.actuallyRender(poseStack, animatable, model, renderType, bufferSource, buffer, isReRender, partialTick, packedLight, packedOverlay, red, green, blue, alpha);
         poseStack.popPose();
-
     }
 }